Foundation implementation of response media field
This is a basic foundations for a responsive media content field to abort
allow content editors to handle responive images. The field can be
configured to have predetermined breakpoints, which will be used to
render a `picture` element on the page. Under the hood the field will
store JSON that represents the media for the different configured
breakpoints. Below is an example of the JSON:

```
[{
    "media": [{
        "breakpoint": 600,
        "path": "/Media/600.jpg"
    }, {
        "breakpoint": 480,
        "path": "/Media/480.jpg"
    }]
}]
```

When rendering the `picture` element, if a configured breakpoint is
not defined the source set will use a resized version of the largest
image.

Currently the admin UI is a `textarea` for the JSON, this needs to be
enhanced to an easy to use UI for selecting media and customising for
different breakpoints.