Add basics for Dictionary Field
Add the basics for the Dictionary Field.

Add a driver for the field itself which currently has edit and update
methods implemented.

Add the ContentField implementing class which has a Dictionary<string,
string> to hold the data.

Add an Edit view model for the edit shape driver which has the Data in
string form for the editor.

Add settings model, driver and view. Currently this just allows setting a
Hint for the field.

Add a view for the Field which currently renders a text field allowing you
to edit the JSON, this will be swapped out for a Vue app allowing managing
a list of name, value properties.

Update Startup with new dependencies.