Add front end pipeline for Dictionary
Add the front end pipeline for the DictionaryField feature.

Add a new entry to the webpack config and convert the existing repsonsive
media entry to a named one using the object approach.

Update the webpack outputs to use the name of the entry as the directory
within wwwroot which we output the admin.js to.

Add a PreBuild event to the csproj on Release configuration which will run
the front end build to ensure this is included in the resulting package.