Add Vue powered editor for DictionaryField
Add a Vue based component for managing the names and values in a
DictionaryField.

Add DictionaryItem model which represents a single entry in the
dictionary, it is implemented as a class so we could extend it with more
fields later if required.

Add a dictionaryEditor Vue component which hooks into the HTML element
provided by the field markup and manages a list of javascript
DictionaryItem models.

Update the markup in the DictionaryField Editor to place the JSON
representation of the dictionary into a hidden field which Vue will pull
the initial value from and then maintain the value as it is updated.

Add a list of items using Vue binding with a vuedraggable wrapper.

Add an Add button which simply tells Vue to add another empty
DictionaryItem to its array which will cause a new row of markup to
appear.

Add remove buttons to each row which will splice the underlying model at
their index and cause the markup to redraw without the row.

Add a handle to the list to highlight the drag and drop functionality.