Add Dictionary Field Max/Min entries settings
Add new settings for the Dictionary Field to allow specifying a Minimum
and Maximum number of entries that can/must be specified.

Update the Driver to check the settings of Min/Max entries against the
specified Data in the Update method. It will add validation errors to the
Data field when there are any.

Update the editor view to show validation errors for the Data field on the
inputs.