Add Dictionary Field Min/Max front end validation
Add front end validation to the Vue component for the Dictionary Field.

Add hidden fields to the Edit shape for Dictionary Field.

Update the Vue component to accept the hidden field min/max values, parse
them and then hide/show various messages based on those numbers. It will
disable the Add button when the max is reached or exceeded.