Fix dictionary field in flows
Fix an issue when the dictionary field is used in a flow by changing how
the default values from the settings are applied. Now the Field has Data
as null by default and the defaults are applied by the Edit Get in the
driver if that field value is null.

Previously this issue was occuring because the code was detecting IsNew by
checking Field.ContentItem.Id which was 0 in the case of a flow widget.