Fix DictionaryField previews
Fix the DictionaryField not updating the live content preview when changes
were made.

Add the content-preview-text class to the input text fields so that typing
will update the preview.

Add a method to the Vue app for updatePreview which will fire the content
preview rendering event on the next text.

Add a watcher to the Vue app which watches dictionaryItems for changes
then calls updatepreview.

Add a change listener to the draggable component which calls
updatepreview.