Add keyboard shortcuts to Dictionary Field
Add some keyboard shortcuts to the Dictionary Field editor.

Add handling for Enter keypress to advance to the next field, adding a new
item if we've reached the end of the list.

Add handling for backspace keypress to go to the previous field if the
field is empty, if we've emptied the name field this will delete the row
if the values field is also empty.

Add handling for up and down arrows to move up and down through the
current column (names/values).