MVP for responsive media field editor
Enhanced the editor experience for managing responsive media. The MVP
allows editors to select an image from the asset library and then
view/define images for the different breakpoints that have been
configured within the field settings.

One thing that will definitely need revisiting is the TypeScript within
the VueJS components to reduce the occurrences of type casting. This
needs a better understanding of the best approach for making VueJS &
TypeScript work together with stronger types.