Fix issue when parsing data value that's been set by recipe
When defining the responsive media field within a recipe, the JSON seems
to be converted from an array to a single item, which was causing the
parsing to throw an error as it's expecting an array. An additional
check to see if it's an array or object is done before deserializing
with the appropriate type.