Fix smallest image not being included on RMF
Fix the smallest breakpoint not being rendered on the front end for the
ResponsiveMediaField by refactoring the loop in GetSourceSets to not skip
the smallest breakpoint and also adjust it to save the last real media
item it saw instead of using the previous source set to prevent ?width
query string parameters building up. Replace the SmallestImageUrl helper
property with a regular property and have GetSourceSets set that after
processing the sources.

Paired with Peter Keating.