Add Render Alias Field
Add a new field called Render Alias Field, this field can render a content
item found via the alias in the specified display type.

Add basic driver, edit view and field model for storing the two string
properties. Default the DisplayType to Detail for simplicity.

Add view models for display and edit.

Add a view for rendering the display mode using Liquid as it has the
simplest markup for this task.

Update the manifest with the new feature and add a startup behind the
feature flag attribute.

Update the version of the module.