Fix issues when OrchardCore.Media.Azure enabled
Fix and issue that would occur when OrchardCore.Media.Azure was enabled
which would mean you could not select images for the responsive media
field.

Add a new utility getUrlFromSelectedMediaItem which will handle the
process of getting the absolute URL for the media item from the selected
media item returned by the media library. Checks if the URL is relative
and prepends the current site origin, if not it will not. The issue before
was that we were always prepending the origin even when the url had
PublicHostname as part of it.