Fix image processing logic on responsive media field
When a user selects multiple images, each one is processed against the
specified breakpoints in the field settings to determine which image
(using it's width) should be associated to each breakpoint.

The logic had an issue though where it would use the first breakpoint
that was lower or equal to the width of the image. This has been changed
to loop over all breakpoints storing any that are lower or equal and
then associating the largest one with the image.