Fix DictionaryField settings not saving
Fix issues that were causing the DictionaryField settings to not save.

Fix Json conversion issues related to the DefaultData which were throwing
exceptions, this was what was causing the settings to not save. Wrap the
JSON conversion in Try/Catch block with logging in case of error.

Fix an error in the dictionary field editor JS on the settings page by
making it handle the min/max hidden fields being missing gracefully
instead of crashing out the editor so you can't do anything.