Add useful extension methods for reading/writing content fields
We've found we regurlarly add extension methods for reading/writing
values on content fields. The extension methods boils the updating of
a value down to a single line instead of the few lines necessary to
set & apply a content field update.

Haven't catered for all fields, only the primary ones we've added to
projects. Will likely add further extension methods for our own content
fields defined in this module.