Add default CSS classes to all fields
Expanding upon [Adam's PR](https://github.com/EtchUK/Etch.OrchardCore.Fields/pull/81),
this introduces a generic approach for adding CSS classes to all the
custom fields provided by this module that render an output. `Classify`
takes the part field definition to generate a CSS class names that
contain a base class and a modifier. The base CSS class is composed of
the content part name and field type. Then the modifier contains the
name of the field on the part. As an example, if you have a content
part named "Foo" that contain a `CodeField` named "Bar", the CSS class
output will be `foo-code foo-code--bar`.